+++ application-domain/src/main/java/io/github/v1serviceapplication/stay/api/StayApi.java
@@ -1,0 +1,14 @@
+package io.github.v1serviceapplication.stay.api;
+
+import io.github.v1serviceapplication.stay.api.dto.response.QueryStayStatusCodeResponse;
+import io.github.v1serviceapplication.stay.api.dto.response.QueryStayStatusResponse;
+import io.github.v1serviceapplication.stay.code.StayStatusCode;
+
+import java.util.UUID;
+
+public interface StayApi {
+    void setDefaultStay(UUID userId);
+    QueryStayStatusCodeResponse queryStayStatusCode();
+    QueryStayStatusResponse queryStayStatus();
+    void applyStay(StayStatusCode status);
+}
